packages/agents/evidence_forge/requirements.txt
gradient-adk>=0.1.0
gradient>=0.1.0
langgraph>=0.2.0
httpx>=0.27.0
psycopg2-binary>=2.9.9
boto3>=1.34.0
reportlab>=4.2.0
